requirements.txt
langchain
langchain-community
langchain-google-genai
gradio
python-dotenv
pydantic 